+++ src/ContainersPlus/Member.hs
@@ -22,9 +22,20 @@ import qualified Data.Set
 import Data.Map  ( Map )
 import Data.Set  ( Set )
 
+-- hashable ----------------------------
+
+import Data.Hashable  ( Hashable )
+
 -- more-unicode ------------------------
 
 import Data.MoreUnicode.Bool  ( 𝔹 )
+
+-- unordered-containers ----------------
+
+import qualified Data.HashMap.Strict
+import qualified Data.HashSet
+
+import Data.HashSet  ( HashSet )
 
 --------------------------------------------------------------------------------
 
@@ -38,19 +49,37 @@ class HasMember α where
   (∉) ∷  MemberType α → α → 𝔹
   (∉) = notMember
 
+--------------------
+
 instance Eq α ⇒ HasMember [α] where
   type MemberType [α] = α
   member    = Data.List.elem
   notMember = Data.List.notElem
+
+--------------------
 
 instance Ord α ⇒ HasMember (Set α) where
   type MemberType (Set α) = α
   member    = Data.Set.member
   notMember = Data.Set.notMember
 
+--------------------
+
 instance Ord α ⇒ HasMember (Map α β) where
   type MemberType (Map α β) = α
   member    = Data.Map.member
   notMember = Data.Map.notMember
+
+--------------------
+
+instance (Eq α, Hashable α) ⇒ HasMember (HashSet α) where
+  type MemberType (HashSet α) = α
+  member    = Data.HashSet.member
+
+--------------------
+
+instance (Eq α, Hashable α) ⇒ HasMember (Data.HashMap.Strict.HashMap α β) where
+  type MemberType (Data.HashMap.Strict.HashMap α β) = α
+  member    = Data.HashMap.Strict.member
 
 -- that's all, folks -----------------------------------------------------------